requirements.txt
# Astro Runtime includes the following pre-installed providers packages: https://docs.astronomer.io/astro/runtime-image-architecture#provider-packages
dbt-core==1.7.9
dbt-duckdb==1.7.3
astronomer-cosmos>=1.0.2
apache-airflow-providers-dbt-cloud[http]
